Move asm_lang into language.
