+++ sway-core/src/asm_generation/mod.rs
@@ -1,4 +1,4 @@
-use crate::asm_lang::{allocated_ops::AllocatedRegister, virtual_register::*};
+use crate::language::asm_lang::{allocated_ops::AllocatedRegister, virtual_register::*};
 
 use std::collections::BTreeSet;
