+++ sway-core/src/language/asm_lang/virtual_immediate.rs
