+++ sway-core/src/language/asm_lang/virtual_ops.rs
