+++ sway-core/src/language/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod asm_lang;
 pub mod parsed;
